requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
psutil
scapy
werkzeug
pymodbus
ipaddress
netaddr
pytest
colorama
dotenv
pyYAML

